upgrade datafusion to 43.0.0

+++ optd-sqlplannertest/tests/basic/verbose.planner.sql
@@ -26,7 +26,7 @@ select count(*) from t1;
 /*
 PhysicalAgg
 ├── aggrs:Agg(Count)
-│   └── [ 1(u8) ]
+│   └── [ 1(i64) ]
 ├── groups: []
 ├── cost: {compute=5000,io=1000}
 ├── stat: {row_cnt=1000}
